app/requirements.txt
fastapi==0.114.0
uvicorn==0.30.0
psycopg2-binary==2.9.9
